# docs/openapi.yaml
openapi: 3.0.3
info:
  title: TunnelBox API
  version: 0.1.0
  description: Local control plane for quickly exposing internal HTTP/HTTPS origins. Quick needs no Cloudflare account or public domain; Private uses WARP and Access without a domain; Public uses a zone hostname, DNS, and Access.
servers:
  - url: http://127.0.0.1:8080
security:
  - sessionCookie: []
paths:
  /healthz:
    get:
      security: []
      summary: Check process health
      responses:
        '200':
          description: Process is alive.
  /readyz:
    get:
      security: []
      summary: Check storage readiness
      responses:
        '200': { description: Storage is ready. }
        '503': { $ref: '#/components/responses/Error' }
  /api/v1/auth/status:
    get:
      security: []
      summary: Check whether an administrator password has been configured
      responses:
        '200':
          description: Authentication state.
          content:
            application/json:
              schema:
                type: object
                required: [initialized]
                properties:
                  initialized: { type: boolean }
  /api/v1/auth/setup:
    post:
      security: []
      summary: Create the administrator password on first use
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AuthRequest' }
      responses:
        '204': { description: Password created and session cookie set. }
        '400': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
  /api/v1/auth/login:
    post:
      security: []
      summary: Log in with the administrator password
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AuthRequest' }
      responses:
        '204': { description: Session cookie set. }
        '401': { $ref: '#/components/responses/Error' }
  /api/v1/auth/logout:
    post:
      summary: Log out and invalidate the current session
      responses:
        '204': { description: Session cookie cleared. }
  /api/v1/integrations/cloudflare:
    put:
      summary: Validate and save Cloudflare credentials
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CloudflareConfigureRequest' }
      responses:
        '200':
          description: Validated integration.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CloudflareStatus' }
        '400': { $ref: '#/components/responses/Error' }
        '502': { $ref: '#/components/responses/Error' }
  /api/v1/integrations/cloudflare/status:
    get:
      summary: Check Cloudflare integration status
      responses:
        '200':
          description: Current status. The token value is never returned.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CloudflareStatus' }
  /api/v1/zones:
    get:
      summary: List zones visible to the configured token
      responses:
        '200':
          description: Zones.
          content:
            application/json:
              schema:
                type: object
                required: [zones]
                properties:
                  zones:
                    type: array
                    items: { $ref: '#/components/schemas/Zone' }
  /api/v1/connectors:
    get:
      summary: List managed cloudflared processes
      responses:
        '200':
          description: Connector status.
          content:
            application/json:
              schema:
                type: object
                required: [connectors]
                properties:
                  connectors:
                    type: array
                    items: { $ref: '#/components/schemas/Connector' }
  /api/v1/services:
    get:
      summary: List services
      responses:
        '200':
          description: Services.
          content:
            application/json:
              schema:
                type: object
                required: [services]
                properties:
                  services:
                    type: array
                    items: { $ref: '#/components/schemas/Service' }
    post:
      summary: Create a service
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ServiceInput' }
      responses:
        '201':
          description: Created service.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Service' }
        '400': { $ref: '#/components/responses/Error' }
  /api/v1/services/{id}:
    parameters:
      - { $ref: '#/components/parameters/ServiceID' }
    get:
      summary: Get a service
      responses:
        '200':
          description: Service.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Service' }
        '404': { $ref: '#/components/responses/Error' }
    patch:
      summary: Update a draft or inactive service
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ServicePatch' }
      responses:
        '200':
          description: Updated service.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Service' }
        '400': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
    delete:
      summary: Delete a service and its owned remote resources
      description: Drafts without remote references are deleted immediately. Services with Cloudflare references start an asynchronous cleanup operation; poll the operation resource until it completes.
      responses:
        '204': { description: Deleted. }
        '202':
          description: Cleanup operation accepted.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Operation' }
        '409': { $ref: '#/components/responses/Error' }
  /api/v1/services/{id}/deploy:
    parameters:
      - { $ref: '#/components/parameters/ServiceID' }
    post:
      summary: Start an asynchronous deployment
      responses:
        '202':
          description: Operation accepted.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Operation' }
        '409': { $ref: '#/components/responses/Error' }
  /api/v1/services/{id}/stop:
    parameters:
      - { $ref: '#/components/parameters/ServiceID' }
    post:
      summary: Start an asynchronous Connector shutdown
      description: Stops the local cloudflared process while retaining Cloudflare resources and service configuration for a later deployment.
      responses:
        '202':
          description: Operation accepted.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Operation' }
        '409': { $ref: '#/components/responses/Error' }
        '502': { $ref: '#/components/responses/Error' }
  /api/v1/operations/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    get:
      summary: Read operation progress
      responses:
        '200':
          description: Operation.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Operation' }
        '404': { $ref: '#/components/responses/Error' }
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: tunnelbox_session

  parameters:
    ServiceID:
      name: id
      in: path
      required: true
      schema: { type: string }
  responses:
    Error:
      description: Stable error response.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorBody' }
  schemas:
    AuthRequest:
      type: object
      required: [password]
      properties:
        password: { type: string, format: password, writeOnly: true, minLength: 8, maxLength: 256 }
    CloudflareConfigureRequest:
      type: object
      required: [account_id, token]
      properties:
        account_id: { type: string }
        zone_id:
          type: string
          description: Required only for Public mode. Leave empty for Private or Quick.
        token: { type: string, format: password, writeOnly: true }
    CloudflareStatus:
      type: object
      required: [configured]
      properties:
        configured: { type: boolean }
        account_id: { type: string }
        zone_id: { type: string }
        token_id: { type: string }
        token_state: { type: string }
        last_error: { type: string }
    Zone:
      type: object
      required: [id, name]
      properties:
        id: { type: string }
        name: { type: string }
    Connector:
      type: object
      required: [service_id, running, healthy]
      properties:
        service_id: { type: string }
        mode: { type: string, enum: [quick, managed] }
        running: { type: boolean }
        healthy: { type: boolean }
        url:
          type: string
          format: uri
          description: Temporary trycloudflare.com URL for a Quick connector, when available.
        message: { type: string }
    ServiceInput:
      type: object
      required: [mode, name, origin_url]
      description: >-
        Mode-specific service input. Quick needs only mode, name and origin_url.
        Private uses hostname as a private IP and requires an Access allow condition.
        Public uses hostname as a zone hostname and requires an Access allow condition.
      properties:
        name: { type: string, maxLength: 120 }
        mode:
          type: string
          enum: [quick, private, public]
        hostname:
          type: string
          description: Private IP for Private mode, or full DNS hostname for Public mode. Omit for Quick.
        origin_url: { type: string, format: uri }
        allow_type:
          type: string
          enum: [email, email_domain]
          description: Required for Private and Public; omit for Quick.
        allow_value:
          type: string
          description: Required for Private and Public; omit for Quick.
    ServicePatch:
      description: All fields are optional for PATCH.
      type: object
      properties:
        name: { type: string, maxLength: 120 }
        mode: { type: string, enum: [quick, private, public] }
        hostname:
          type: string
          description: Private IP for Private mode, or full DNS hostname for Public mode.
        origin_url: { type: string, format: uri }
        allow_type: { type: string, enum: [email, email_domain] }
        allow_value: { type: string }
    Service:
      allOf:
        - $ref: '#/components/schemas/ServiceInput'
        - type: object
          required: [id, state, created_at, updated_at]
          properties:
            id: { type: string }
            state: { type: string, enum: [draft, deploying, stopping, active, stopped, error] }
            tunnel_id: { type: string }
            private_route_id: { type: string }
            dns_record_id: { type: string }
            access_application_id: { type: string }
            access_policy_id: { type: string }
            public_url:
              type: string
              format: uri
              description: Generated Quick URL, when the service is active.
            created_at: { type: string, format: date-time }
            updated_at: { type: string, format: date-time }
    Operation:
      type: object
      required: [operation_id, service_id, kind, status, attempts, created_at, updated_at]
      properties:
        operation_id: { type: string }
        service_id: { type: string }
        kind: { type: string }
        status: { type: string, enum: [pending, running, succeeded, failed, unknown] }
        current_step: { type: string }
        attempts: { type: integer }
        error_code: { type: string }
        error_message: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        started_at: { type: string, format: date-time, nullable: true }
        finished_at: { type: string, format: date-time, nullable: true }
    ErrorBody:
      type: object
      required: [code, message, request_id]
      properties:
        code: { type: string }
        message: { type: string }
        request_id: { type: string }