# internal/httpapi/router.go
package httpapi

import (
	"context"
	"errors"
	"log/slog"
	"net/http"
	"path/filepath"
	"strings"
	"time"

	"github.com/gin-gonic/gin"
	"github.com/sparklyi/tunnelbox/internal/auth"
	"github.com/sparklyi/tunnelbox/internal/operation"
	"github.com/sparklyi/tunnelbox/internal/provision"
	"github.com/sparklyi/tunnelbox/internal/service"
)

// ServiceActions is the small API contract needed by the HTTP handlers.
type ServiceActions interface {
	List(context.Context) ([]service.Service, error)
	Get(context.Context, string) (service.Service, error)
	Create(context.Context, service.CreateInput) (service.Service, error)
	Update(context.Context, string, service.UpdateInput) (service.Service, error)
	Delete(context.Context, string) error
}

// OperationReader is kept separate so read-only operation endpoints do not
// depend on the operation runner implementation.
type OperationReader interface {
	Get(context.Context, string) (operation.Operation, error)
}

// ServiceDeployer starts an asynchronous deployment for one service.
type ServiceDeployer interface {
	Deploy(context.Context, string) (operation.Operation, error)
}

// ServiceStopper starts an asynchronous shutdown of one service's Connector.
type ServiceStopper interface {
	Stop(context.Context, string) (operation.Operation, error)
}

// ServiceDeleter removes a service immediately or starts its asynchronous
// remote cleanup. A zero operation means the local record was deleted.
type ServiceDeleter interface {
	Delete(context.Context, string) (operation.Operation, error)
}

type CloudflareIntegration interface {
	Configure(context.Context, provision.CloudflareConfigureInput) (provision.CloudflareIntegrationStatus, error)
	Status(context.Context) (provision.CloudflareIntegrationStatus, error)
	Zones(context.Context) ([]provision.Zone, error)
}

type ConnectorLister interface {
	List(context.Context) ([]provision.ConnectorStatus, error)
}

type Dependencies struct {
	Services      ServiceActions
	Operations    OperationReader
	Deployer      ServiceDeployer
	Stopper       ServiceStopper
	Deleter       ServiceDeleter
	Cloudflare    CloudflareIntegration
	Connectors    ConnectorLister
	Auth          *auth.Manager
	SecureCookies bool
	Logger        *slog.Logger
	Readiness     func(context.Context) error
	WebDir        string
}

func NewRouter(deps Dependencies) (*gin.Engine, error) {
	if deps.Services == nil {
		return nil, errors.New("http api requires service actions")
	}
	if deps.Operations == nil {
		return nil, errors.New("http api requires operation reader")
	}
	if deps.Logger == nil {
		deps.Logger = slog.Default()
	}

	router := gin.New()
	router.Use(requestIDMiddleware(), recoveryMiddleware(deps.Logger), loggingMiddleware(deps.Logger), errorMiddleware())
	router.Use(authMiddleware(deps.Auth))

	router.GET("/healthz", healthHandler)
	router.GET("/readyz", func(c *gin.Context) {
		if deps.Readiness != nil {
			if err := deps.Readiness(c.Request.Context()); err != nil {
				writeError(c, http.StatusServiceUnavailable, "not_ready", "tunnelbox is not ready")
				return
			}
		}
		c.JSON(http.StatusOK, gin.H{"status": "ok"})
	})

	api := router.Group("/api/v1")
	authLimiter := newAuthAttemptLimiter(5, 5*time.Minute)
	api.GET("/auth/status", authStatusHandler(deps.Auth))
	api.POST("/auth/setup", authSetupHandler(deps.Auth, deps.SecureCookies, authLimiter))
	api.POST("/auth/login", authLoginHandler(deps.Auth, deps.SecureCookies, authLimiter))
	api.POST("/auth/logout", authLogoutHandler(deps.Auth, deps.SecureCookies))
	api.PUT("/integrations/cloudflare", configureCloudflareHandler(deps.Cloudflare))
	api.GET("/integrations/cloudflare/status", cloudflareStatusHandler(deps.Cloudflare))
	api.GET("/zones", listZonesHandler(deps.Cloudflare))
	api.GET("/connectors", listConnectorsHandler(deps.Connectors))
	api.GET("/services", listServicesHandler(deps.Services))
	api.POST("/services", createServiceHandler(deps.Services))
	api.GET("/services/:id", getServiceHandler(deps.Services))
	api.PATCH("/services/:id", updateServiceHandler(deps.Services))
	api.DELETE("/services/:id", deleteServiceHandler(deps.Services, deps.Deleter))
	api.POST("/services/:id/deploy", deployServiceHandler(deps.Deployer))
	api.POST("/services/:id/stop", stopServiceHandler(deps.Stopper))
	api.GET("/operations/:id", getOperationHandler(deps.Operations))
	if webDir := strings.TrimSpace(deps.WebDir); webDir != "" {
		webDir = filepath.Clean(webDir)
		router.GET("/", func(c *gin.Context) { c.File(filepath.Join(webDir, "index.html")) })
		router.Static("/assets", filepath.Join(webDir, "assets"))
	}
	return router, nil
}

func healthHandler(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{"status": "ok"})
}

func contextForOperation(c *gin.Context) context.Context {
	// A disconnected HTTP client must not cancel an operation already accepted.
	return context.WithoutCancel(c.Request.Context())
}